S3ConfigurationRepositoryBuilder
This commit introduces an S3ConfigurationRepositoryBuilder and reverts making the constructor in S3FileLoader public. Also, the previously public constructors of S3ConfigurationRepository are marked deprecated. Instead, there is a package-private constructor that is used by the new builder class. This allows fully customizing the S3 configuration repository, e.g. for testing, but still provides meaningful defaults for clients who do not require this customization. It also prevents leaking of implementation details (S3FileLoader) to the client introduced in the previous commit.

Further changes:
- add Lukas Wilhelm to MAINTAINERS
- cleanup in test class