Enable structured types as config values
This commit enables structured types as config values. That is, something like
```
@BaiganConfig
public interface SomeComplexConfiguration {
    SomeConfigObject someConfig();
}
```
can now be deserialized, e.g. from a config file with content
```
[{ "alias": "some.complex.configuration.some.config", "defaultValue": {"config_key":"a value"}}]
```

How it works is that the repositories have to be constructed with a class `BaiganConfigClasses`, a container for all interfaces annotated with `@BaiganConfig`. The repository then deserializes first to `JsonNode`, then deserializes this to the correct class for each config key.
The `BaiganConfigClasses` is provided as a bean. It is constructed by the `ContextAwareConfigurationMethodInvocationHandler` when creating the proxies for the `BaiganConfig` interfaces.

Beyond this bean, users of the library do not have to add any extra configuration to enable structured config types. They only have to make sure that the `ObjectMapper` can deserialize the classes correctly. To this end, the `ObjectMapper` can now be injected into the repositories.

The change is backwards compatible with respect to config files. All old config files can be processes as before. This is covered in an End2End test. It is not generally backwards compatible regarding the code as deprecated constructors for the `S3ConfigurationRepository` were removed and all constructors for repositories now require `BaiganConfigClasses` to be passed.