Simplify ContextAwareConfigurationMethodInvocationHandler
As the repositories now return the type checking and constructor logic can now simplified.

Primitive types would in principle work. However, if no config can be found, then trying to return null for a primitive return type of a configuration would cause a NullPointerException. To make sure users do not declare primitives as return types, the BaiganConfigClasses throws and exception if there are primitive return types, causing the starting of the ApplicationContext to fail.