remove AbstractConfigurationRepository and use composition instead of inheritance
The inheritance implied by the AbstractConfigurationRepository makes it harder to test its logic and that of the repositories that use it. This commit changes it to composition, turning the abstract class into a service that is used by the repositories. The EtcdConfigurationRepository did not actually use any of the functionality of the abstract class, so it does not have the new service as component.

Also, this commit increases the test coverage.