process generics in config types correctly
This commit adds support for handling generics in the retunr type of a config method. Instead of parsing `Map<UUID, List<SomeConfigObject>>` to a `Map<String, List<Map<...>>>`, it is now correctly processed to the expected type.

This works by using `Type` instead of `Class<?>` in `BaiganConfigClasses` and actually using the `AnnotatedType` instead of the plain type when constructing the bean in `ConfigurationBeanRegistrar`.