introduce RepositoryFactory
Previously, the repositories worked only when created as Spring beans due to the injection of the `ConfigurationParser`. This was at odds with their constructors being accessible. In particular, this broke usage of the ChainedConfigurationRepository if the repositories passed to it were not created as beans themselves.

This is solved in this commit by adding a `RepositoryFactory` bean through which all repositories are created.

Also, the class `ConfigTypeProvider` is removed as it was an unnecessary layer just wrapping `BaiganConfigClasses`.